wip: centralize setLocalRobot as duplicate
