+++ robot/impl/resource_manager_test.go
@@ -1845,20 +1845,3 @@ func managerForDummyRobot(t *testing.T, robot robot.Robot) *resourceManager {
 	}
 	return manager
 }
-
-func setupLocalRobot(
-	t *testing.T,
-	ctx context.Context,
-	cfg *config.Config,
-	logger logging.Logger,
-) robot.LocalRobot {
-	t.Helper()
-
-	r, err := New(ctx, cfg, logger)
-	test.That(t, err, test.ShouldBeNil)
-	test.That(t, r, test.ShouldNotBeNil)
-	t.Cleanup(func() {
-		test.That(t, r.Close(ctx), test.ShouldBeNil)
-	})
-	return r
-}
